20240929_xjh_Adjust the relevant layout
